Updated packages; improved handling of bridge database sync
