+++ controller/apps/rides-examiner/package.json
@@ -16,15 +16,15 @@
 	"dependencies": {
 		"@helperkits/logger": "20240703.1726.24",
 		"@helperkits/timer": "20240627.34.23",
-		"@tmlmobilidade/interfaces": "20250612.1741.30",
+		"@tmlmobilidade/interfaces": "20250626.1107.33",
 		"@tmlmobilidade/sae-controller-pckg-utils": "*",
-		"@tmlmobilidade/utils": "20250612.1741.30",
+		"@tmlmobilidade/utils": "20250626.1107.33",
 		"@turf/turf": "7.2.0"
 	},
 	"devDependencies": {
-		"@tmlmobilidade/tsconfig": "20250612.1741.30",
-		"@tmlmobilidade/types": "20250612.1741.30",
-		"@types/node": "24.0.1",
+		"@tmlmobilidade/tsconfig": "20250626.1107.33",
+		"@tmlmobilidade/types": "20250626.1107.33",
+		"@types/node": "24.0.4",
 		"resolve-tspaths": "0.8.23",
 		"tsx": "4.20.3",
 		"typescript": "5.8.3"
